+++ extension/lib/userUpdate.js
@@ -30,7 +30,8 @@ module.exports = async (context, { firstName, lastName, customAttributes }) => {
     customerRepo = BigCommerceCustomerRepository.create(
       context.config.clientId,
       context.config.accessToken,
-      context.config.storeHash
+      context.config.storeHash,
+      context.log
     )
   }