+++ extension/lib/login.js
@@ -1,7 +1,7 @@
 const InvalidCredentialsError = require('./shopgate/customer/errors/InvalidCredentialsError')
 const BigCommerceCustomerRepository = require('./bigcommerce/CustomerRepository')
 const { decorateError } = require('./shopgate/logDecorator')
-const getCustomer = require('./shopgate/customer/get')
+const { getCustomer } = require('./shopgate/customer/get')
 const UnknownError = require('./shopgate/customer/errors/UnknownError')
 const UserNotFoundError = require('./shopgate/customer/errors/UserNotFoundError')
 
@@ -17,7 +17,8 @@ module.exports = async (context, input) => {
     customerRepo = BigCommerceCustomerRepository.create(
       context.config.clientId,
       context.config.accessToken,
-      context.config.storeHash
+      context.config.storeHash,
+      context.log
     )
   }