+++ extension/test/integration/lib/bigcommerce/CustomerRepository-test.js
@@ -9,13 +9,15 @@ const { sleep } = require('../../../util/sleep')
 describe('BigCommerceCustomerRepository', () => {
   /** @var {BigCommerceCustomerRepository} */
   let repo
+  /** @type Logger */
+  const logger = { debug: () => {} }
 
   beforeEach(() => {
-    repo = BigCommerceCustomerRepository.create(clientId, accessToken, storeHash)
+    repo = BigCommerceCustomerRepository.create(clientId, accessToken, storeHash, logger)
   })
 
   describe('update()', () => {
-    const customerId = 317
+    const customerId = 374
 
     it('should update first name', async () => {
       await repo.update(customerId, {