+++ CHANGELOG.md
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+[2.0.0] - 2019-02-20
 ### Added
 - basic user data management
+- user data can now be expired, time to live is configurable
+- logging for all API calls
+## Changed
+- Frontend migration to PWA6.
 ### Fixed
 - handling user not found case during login process
 
@@ -40,7 +46,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - login pipeline now calls an additional step at the end (from the bigcommerce cart extension) that merges the anonymous cart into the logged in cart
 
-[Unreleased]: https://github.com/shopgate/ext-bigcommerce-user/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/shopgate/ext-bigcommerce-user/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/shopgate/ext-bigcommerce-user/compare/v0.4.0...v2.0.0
 [0.4.1]: https://github.com/shopgate/ext-bigcommerce-user/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/shopgate/ext-bigcommerce-user/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/shopgate/ext-bigcommerce-user/compare/v0.2.0...v0.3.0